Layout Home Check
Added 'home' prop to check for in Layout.js. If 'home' is present, then show 'add' button at bottom of screen to add a new recipe. If 'home' does not exist, then don't include add button to open side form drawer. This is to prevent add button from showing up on About page and future pages that do not make sense to have that option.